+++ modules/java/main/src/main/java/com/mooltiverse/oss/nyx/template/Functions.java
@@ -8,10 +8,8 @@ import java.io.StringWriter;
 import java.text.SimpleDateFormat;
 import java.time.ZoneOffset;
 import java.util.Date;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
-import java.util.function.Function;
 import java.util.TimeZone;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -31,38 +29,6 @@ class Functions {
      * The private logger instance
      */
     private static final Logger logger = LoggerFactory.getLogger(Functions.class);
-
-    /**
-     * The map of all available functions, where keys are function names to be used in various contexts,
-     * and values are function implementations.
-     */
-    public static Map<String,Function<String,String>> FUNCTIONS = new HashMap<String,Function<String,String>>(){
-        {
-            put(Lower.NAME,                   new Lower());
-            put(Upper.NAME,                   new Upper());
-            put(Trim.NAME,                    new Trim());
-            put(First.NAME,                   new First());
-            put(FirstLower.NAME,              new FirstLower());
-            put(FirstUpper.NAME,              new FirstUpper());
-            put(Last.NAME,                    new Last());
-            put(LastLower.NAME,               new LastLower());
-            put(LastUpper.NAME,               new LastUpper());
-            put(Sanitize.NAME,                new Sanitize());
-            put(SanitizeLower.NAME,           new SanitizeLower());
-            put(SanitizeUpper.NAME,           new SanitizeUpper());
-            put(Short5.NAME,                  new Short5());
-            put(Short6.NAME,                  new Short6());
-            put(Short7.NAME,                  new Short7());
-            put(TimestampISO8601.NAME,        new TimestampISO8601());
-            put(TimestampYYYYMMDDHHMMSS.NAME, new TimestampYYYYMMDDHHMMSS());
-
-            put(EnvironmentUser.NAME,         new EnvironmentUser());
-            put(EnvironmentVariable.NAME,     new EnvironmentVariable());
-
-            put(FileContent.NAME,             new FileContent());
-            put(FileExists.NAME,              new FileExists());
-        }
-    };
 
     /**
      * Default constructor is private on purpose.
@@ -100,30 +66,83 @@ class Functions {
 
         registry.registerHelper(FileContent.NAME,             new FileContent());
         registry.registerHelper(FileExists.NAME,              new FileExists());
+
+        registry.registerHelper(Capture.NAME,                 new Capture());
+        registry.registerHelper(CutLeft.NAME,                 new CutLeft());
+        registry.registerHelper(CutRight.NAME,                new CutRight());
+        registry.registerHelper(Replace.NAME,                 new Replace());
+        registry.registerHelper(TimeFormat.NAME,              new TimeFormat());
     }
 
     /**
-     * This class provides basic features for functions.
+     * This is the superclass for all template functions.
      * 
      * In particular this class offers the {@link Helper} implementation as an adapter method
      * used by Handlebars to invoke the actual business method of the function.
      */
-    static abstract class AbstractFunction implements Function<String,String>, Helper<Object> {
+    static abstract class AbstractFunction implements Helper<Object> {
+    }
+
+    /**
+     * This class provides basic features for functions with options (parameters).
+     * This kind of function has a name (the function name), one input which is the text from the template,
+     * and one or more parameters.
+     */
+    static abstract class AbstractParametricFunction extends AbstractFunction {
         /**
-         * {@inheritDoc}}
+         * {@inheritDoc}
          */
         @Override
         public Object apply(Object context, Options options)
             throws IOException {
-            // The docs are poorly documented by all examples show that options.fn(this) is the only way to have the block resolved
+            // The docs are poorly documented by all examples show that options.fn(this)
+            // is the only way to have the block resolved
+            return (Objects.isNull(options) ? "" : apply(options.fn(this).toString(), options.hash));
+        }
+
+        /**
+         * This method must be implemented by subclasses and will return the function specific output,
+         * provided the input and options.
+         * 
+         * @param input the function input
+         * @param options the function options
+         * 
+         * @return the function output
+         */
+        public abstract String apply(String input, Map<String,Object> options);
+    }
+
+    /**
+     * This class provides basic features for functions without options (parameters).
+     * This kind of function only has a name (the function name) and one input which is the text from the template.
+     */
+    static abstract class AbstractSimpleFunction extends AbstractFunction {
+        /**
+         * {@inheritDoc}
+         */
+        @Override
+        public Object apply(Object context, Options options)
+            throws IOException {
+            // The docs are poorly documented by all examples show that options.fn(this)
+            // is the only way to have the block resolved
             return (Objects.isNull(options) ? "" : apply(options.fn(this).toString()));
         }
+
+        /**
+         * This method must be implemented by subclasses and will return the function specific output,
+         * provided the input.
+         * 
+         * @param input the function input
+         * 
+         * @return the function output
+         */
+        public abstract String apply(String input);
     }
 
     /**
      * This function returns the lower case representation of the input string.
      */
-    public static class Lower extends AbstractFunction {
+    public static class Lower extends AbstractSimpleFunction {
         /**
          * The name to use for this function.
          */
@@ -136,6 +155,7 @@ class Functions {
          * 
          * @return the lower case representation of the input string.
          */ 
+        @Override
         public String apply(String input) {
             if (Objects.isNull(input))
                 return "";
@@ -147,7 +167,7 @@ class Functions {
     /**
      * This function returns the trimmed case representation of the input string.
      */
-    public static class Trim extends AbstractFunction {
+    public static class Trim extends AbstractSimpleFunction {
         /**
          * The name to use for this function.
          */
@@ -159,7 +179,8 @@ class Functions {
          * @param input the input string. If {@code null} an empty string is returned.
          * 
          * @return the trimmed case representation of the input string.
-         */ 
+         */
+        @Override
         public String apply(String input) {
             if (Objects.isNull(input))
                 return "";
@@ -171,7 +192,7 @@ class Functions {
     /**
      * This function returns the upper case representation of the input string.
      */
-    public static class Upper extends AbstractFunction {
+    public static class Upper extends AbstractSimpleFunction {
         /**
          * The name to use for this function.
          */
@@ -183,7 +204,8 @@ class Functions {
          * @param input the input string. If {@code null} an empty string is returned.
          * 
          * @return the upper case representation of the input string.
-         */ 
+         */
+        @Override
         public String apply(String input) {
             if (Objects.isNull(input))
                 return "";
@@ -196,7 +218,7 @@ class Functions {
      * This function returns the input string with everything from the first occurrence of a character other
      * than letters and positive digits discarded.
      */
-    public static class First extends AbstractFunction {
+    public static class First extends AbstractSimpleFunction {
         /**
          * The name to use for this function.
          */
@@ -210,7 +232,8 @@ class Functions {
          * 
          * @return the input string with everything from the first occurrence of a character other
          * than letters and positive digits discarded.
-         */ 
+         */
+        @Override
         public String apply(String input) {
             if (Objects.isNull(input))
                 return "";
@@ -224,7 +247,7 @@ class Functions {
      * This function returns the input string with everything from the first occurrence of a character other
      * than letters and positive digits discarded and the remainder transformed to lower case.
      */
-    public static class FirstLower extends AbstractFunction {
+    public static class FirstLower extends AbstractSimpleFunction {
         /**
          * The name to use for this function.
          */
@@ -238,7 +261,8 @@ class Functions {
          * 
          * @return the input string with everything other than letters and positive digits discarded
          * and the remainder transformed to lower case.
-         */ 
+         */
+        @Override
         public String apply(String input) {
             if (Objects.isNull(input))
                 return "";
@@ -251,7 +275,7 @@ class Functions {
      * This function returns the input string with everything from the first occurrence of a character other
      * than letters and positive digits discarded and the remainder transformed to upper case.
      */
-    public static class FirstUpper extends AbstractFunction {
+    public static class FirstUpper extends AbstractSimpleFunction {
         /**
          * The name to use for this function.
          */
@@ -265,7 +289,8 @@ class Functions {
          * 
          * @return the input string with everything other than letters and positive digits discarded
          * and the remainder transformed to upper case.
-         */ 
+         */
+        @Override
         public String apply(String input) {
             if (Objects.isNull(input))
                 return "";
@@ -278,7 +303,7 @@ class Functions {
      * This function returns the last part of the input string that does not contains characters other than
      * letters and positive digits.
      */
-    public static class Last extends AbstractFunction {
+    public static class Last extends AbstractSimpleFunction {
         /**
          * The name to use for this function.
          */
@@ -292,7 +317,8 @@ class Functions {
          * 
          * @return the last part of the input string that does not contains characters other than.
          * letters and positive digits.
-         */ 
+         */
+        @Override
         public String apply(String input) {
             if (Objects.isNull(input))
                 return "";
@@ -306,7 +332,7 @@ class Functions {
      * This function returns the last part of the input string that does not contains characters other than
      * letters and positive digits and the remainder transformed to lower case.
      */
-    public static class LastLower extends AbstractFunction {
+    public static class LastLower extends AbstractSimpleFunction {
         /**
          * The name to use for this function.
          */
@@ -320,7 +346,8 @@ class Functions {
          * 
          * @return the last part of the input string that does not contains characters other than
          * letters and positive digits and the remainder transformed to lower case.
-         */ 
+         */
+        @Override
         public String apply(String input) {
             if (Objects.isNull(input))
                 return "";
@@ -333,7 +360,7 @@ class Functions {
      * This function returns the last part of the input string that does not contains characters other than
      * letters and positive digits and the remainder transformed to upper case.
      */
-    public static class LastUpper extends AbstractFunction {
+    public static class LastUpper extends AbstractSimpleFunction {
         /**
          * The name to use for this function.
          */
@@ -347,7 +374,8 @@ class Functions {
          * 
          * @return the last part of the input string that does not contains characters other than
          * letters and positive digits and the remainder transformed to upper case.
-         */ 
+         */
+        @Override
         public String apply(String input) {
             if (Objects.isNull(input))
                 return "";
@@ -359,7 +387,7 @@ class Functions {
     /**
      * This function returns the input string with everything other than letters and positive digits discarded.
      */
-    public static class Sanitize extends AbstractFunction {
+    public static class Sanitize extends AbstractSimpleFunction {
         /**
          * The name to use for this function.
          */
@@ -371,7 +399,8 @@ class Functions {
          * @param input the input string. If {@code null} an empty string is returned.
          * 
          * @return the input string with everything other than letters and positive digits discarded.
-         */ 
+         */
+        @Override
         public String apply(String input) {
             if (Objects.isNull(input))
                 return "";
@@ -384,7 +413,7 @@ class Functions {
      * This function returns the input string with everything other than letters and positive digits discarded
      * and the remainder transformed to lower case.
      */
-    public static class SanitizeLower extends AbstractFunction {
+    public static class SanitizeLower extends AbstractSimpleFunction {
         /**
          * The name to use for this function.
          */
@@ -398,7 +427,8 @@ class Functions {
          * 
          * @return the input string with everything other than letters and positive digits discarded
          * and the remainder transformed to lower case.
-         */ 
+         */
+        @Override
         public String apply(String input) {
             if (Objects.isNull(input))
                 return "";
@@ -411,7 +441,7 @@ class Functions {
      * This function returns the input string with everything other than letters and positive digits discarded
      * and the remainder transformed to upper case.
      */
-    public static class SanitizeUpper extends AbstractFunction {
+    public static class SanitizeUpper extends AbstractSimpleFunction {
         /**
          * The name to use for this function.
          */
@@ -425,7 +455,8 @@ class Functions {
          * 
          * @return the input string with everything other than letters and positive digits discarded
          * and the remainder transformed to upper case.
-         */ 
+         */
+        @Override
         public String apply(String input) {
             if (Objects.isNull(input))
                 return "";
@@ -438,7 +469,7 @@ class Functions {
      * This function returns the first 5 characters of the input string, if it's longer than 5 characters,
      * otherwise returns the input string.
      */
-    public static class Short5 extends AbstractFunction {
+    public static class Short5 extends AbstractSimpleFunction {
         /**
          * The name to use for this function.
          */
@@ -446,13 +477,14 @@ class Functions {
 
         /**
          * This method returns the first 5 characters of the input string, if it's longer than 5 characters,
-         * otherwise returns the input string or the empty string if the input is {code null}.
+         * otherwise returns the input string or the empty string if the input is {@code null}.
          * 
          * @param input the input string. If {@code null} an empty string is returned.
          * 
          * @return the first 5 characters of the input string, if it's longer than 5 characters,
          * otherwise returns the input string.
-         */ 
+         */
+        @Override
         public String apply(String input) {
             if (Objects.isNull(input))
                 return "";
@@ -465,7 +497,7 @@ class Functions {
      * This function returns the first 6 characters of the input string, if it's longer than 6 characters,
      * otherwise returns the input string.
      */
-    public static class Short6 extends AbstractFunction {
+    public static class Short6 extends AbstractSimpleFunction {
         /**
          * The name to use for this function.
          */
@@ -473,13 +505,14 @@ class Functions {
 
         /**
          * This method returns the first 6 characters of the input string, if it's longer than 6 characters,
-         * otherwise returns the input string or the empty string if the input is {code null}.
+         * otherwise returns the input string or the empty string if the input is {@code null}.
          * 
          * @param input the input string. If {@code null} an empty string is returned.
          * 
          * @return the first 6 characters of the input string, if it's longer than 6 characters,
          * otherwise returns the input string.
-         */ 
+         */
+        @Override
         public String apply(String input) {
             if (Objects.isNull(input))
                 return "";
@@ -492,7 +525,7 @@ class Functions {
      * This function returns the first 7 characters of the input string, if it's longer than 7 characters,
      * otherwise returns the input string.
      */
-    public static class Short7 extends AbstractFunction {
+    public static class Short7 extends AbstractSimpleFunction {
         /**
          * The name to use for this function.
          */
@@ -500,13 +533,14 @@ class Functions {
 
         /**
          * This method returns the first 7 characters of the input string, if it's longer than 7 characters,
-         * otherwise returns the input string or the empty string if the input is {code null}.
+         * otherwise returns the input string or the empty string if the input is {@code null}.
          * 
          * @param input the input string. If {@code null} an empty string is returned.
          * 
          * @return the first 7 characters of the input string, if it's longer than 7 characters,
          * otherwise returns the input string.
-         */ 
+         */
+        @Override
         public String apply(String input) {
             if (Objects.isNull(input))
                 return "";
@@ -520,7 +554,7 @@ class Functions {
      * <a href="https://www.unixtimestamp.com/">unix format</a> and returns it formatted as
      * <a href="https://en.wikipedia.org/wiki/ISO_8601">ISO 8601</a> UTC timestamp.
      */
-    public static class TimestampISO8601 extends AbstractFunction {
+    public static class TimestampISO8601 extends AbstractSimpleFunction {
         /**
          * The name to use for this function.
          */
@@ -537,7 +571,8 @@ class Functions {
          * 
          * @return the ISO 8601 representation of the input timestamp, or an empty string if parsing the input
          * fails for any reson.
-         */ 
+         */
+        @Override
         public String apply(String input) {
             if (Objects.isNull(input))
                 return "";
@@ -560,7 +595,7 @@ class Functions {
      * <a href="https://www.unixtimestamp.com/">unix format</a> and returns it formatted as
      * {@code YYYYMMDDHHMMSS} UTC.
      */
-    public static class TimestampYYYYMMDDHHMMSS extends AbstractFunction {
+    public static class TimestampYYYYMMDDHHMMSS extends AbstractSimpleFunction {
         /**
          * The name to use for this function.
          */
@@ -577,7 +612,8 @@ class Functions {
          * 
          * @return the ISO 8601 representation of the input timestamp, or an empty string if parsing the input
          * fails for any reson.
-         */ 
+         */
+        @Override
         public String apply(String input) {
             if (Objects.isNull(input))
                 return "";
@@ -598,7 +634,7 @@ class Functions {
     /**
      * This function returns the value of the environment variable with the given name, if any.
      */
-    public static class EnvironmentVariable extends AbstractFunction {
+    public static class EnvironmentVariable extends AbstractSimpleFunction {
         /**
          * The name to use for this function.
          */
@@ -610,7 +646,8 @@ class Functions {
          * @param name the name of the requested variable. If {@code null} an empty string is returned.
          * 
          * @return the value of the environment variable with the given name, if any.
-         */ 
+         */
+        @Override
         public String apply(String name) {
             if (Objects.isNull(name))
                 return "";
@@ -623,7 +660,7 @@ class Functions {
     /**
      * This function returns the current user name. The input parameter is ignored.
      */
-    public static class EnvironmentUser extends AbstractFunction {
+    public static class EnvironmentUser extends AbstractSimpleFunction {
         /**
          * The name to use for this function.
          */
@@ -635,7 +672,8 @@ class Functions {
          * @param name ignored.
          * 
          * @return the current user name.
-         */ 
+         */
+        @Override
         public String apply(String name) {
             return System.getProperty("user.name");
         }
@@ -644,7 +682,7 @@ class Functions {
     /**
      * This function returns the content of the given file, if any.
      */
-    public static class FileContent extends AbstractFunction {
+    public static class FileContent extends AbstractSimpleFunction {
         /**
          * The name to use for this function.
          */
@@ -657,7 +695,8 @@ class Functions {
          * against the system current directory (configured directories are ignored here).
          * 
          * @return the content of the given file, if any.
-         */ 
+         */
+        @Override
         public String apply(String path) {
             if (Objects.isNull(path))
                 return "";
@@ -683,7 +722,7 @@ class Functions {
     /**
      * This function returns {@code true} if the file with the given name exists, {@code false} otherwise.
      */
-    public static class FileExists extends AbstractFunction {
+    public static class FileExists extends AbstractSimpleFunction {
         /**
          * The name to use for this function.
          */
@@ -696,7 +735,8 @@ class Functions {
          * current directory (configured directories are ignored here).
          * 
          * @return {@code true} if the file with the given name exists, {@code false} otherwise.
-         */ 
+         */
+        @Override
         public String apply(String path) {
             if (Objects.isNull(path))
                 return Boolean.FALSE.toString();
@@ -704,6 +744,260 @@ class Functions {
             java.io.File f = new java.io.File(path);
             logger.trace(TEMPLATE, "Checking whether file '{}' ('{}') exists", path, f.getAbsolutePath());
             return Boolean.toString(f.exists() && f.isFile());
+        }
+    }
+
+    /**
+     * This function uses a given regular expression to match the string input and uses the given (named) group
+     * value as a return value.
+     */
+    public static class Capture extends AbstractParametricFunction {
+        /**
+         * The name to use for this function.
+         */
+        public static final String NAME = "capture";
+
+        /**
+         * This method uses a given regular expression to match the string input and uses the given (named) group
+         * value as a return value. The {@code expression} option gives the regular expression to use, while the
+         * {@code group} option tells which group to return (if it's an integer it will be interpreded as a group index,
+         * otherwise a group name). Remember that group {@code 0} returns the whole string. If the regular expression
+         * doesn't even match the input string then an empty string is returned.
+         * 
+         * @param input the function input
+         * @param options the function options, where the option named {@code expression} gives the regular expression to use,
+         * while the {@code group} option tells which group to return (if it's an integer it will be interpreded as a group index,
+         * otherwise a group name)
+         * 
+         * @return the function output
+         */
+        @Override
+        public String apply(String input, Map<String,Object> options) {
+            if (Objects.isNull(input))
+                return "";
+            if (options.containsKey("expression")) {
+                String expression = options.get("expression").toString();
+                Matcher m = Pattern.compile(expression).matcher(input);
+                m.find();
+
+                if (options.containsKey("group")) {
+                    if (options.containsKey("group")) {
+                        String groupName = options.get("group").toString();
+                        try {
+                            int groupIndex = Integer.parseUnsignedInt(groupName);
+                            // The 'group' is an index
+                            try {
+                                return m.group(groupIndex);
+                            }
+                            catch (IllegalStateException ise) {
+                                logger.debug(TEMPLATE, "The '{}' expression doesn't match '{}'", expression, input);
+                                return "";
+                            }
+                            catch (IndexOutOfBoundsException ioobe) {
+                                logger.debug(TEMPLATE, "The '{}' expression doesn't return any group with index '{}' from string '{}'", expression, groupIndex, input);
+                                return "";
+                            }
+                        }
+                        catch (NumberFormatException nfe) {
+                            // The 'group' is a name
+                            try {
+                                return m.group(groupName);
+                            }
+                            catch (IllegalStateException ise) {
+                                logger.debug(TEMPLATE, "The '{}' expression doesn't match '{}'", expression, input);
+                                return "";
+                            }
+                            catch (IllegalArgumentException iae) {
+                                logger.debug(TEMPLATE, "The '{}' expression doesn't return any group with name '{}' from string '{}'", expression, groupName, input);
+                                return "";
+                            }
+                        }
+                    }
+                    else return input;
+                }
+                else return input;
+            }
+            else return input;
+        }
+    }
+
+    /**
+     * This function returns the last {@code N} characters of the input string, where {@code N} is the {@code length} option.
+     * If the input string is shorter or the same length than the parameter, the whole input string is returned unchanged.
+     * If the input is {@code null} an empty string is returned.
+     */
+    public static class CutLeft extends AbstractParametricFunction {
+        /**
+         * The name to use for this function.
+         */
+        public static final String NAME = "cutLeft";
+
+        /**
+         * This method returns the last {@code N} characters of the input string, where {@code N} is the {@code length} option.
+         * If the input string is shorter or the same length than the parameter, the whole input string is returned unchanged.
+         * If the input is {@code null} an empty string is returned.
+         * 
+         * @param input the function input
+         * @param options the function options, where the option named {@code length} determines the maximum length of the
+         * returned string. If there is no such option the same string given as input is returned
+         * 
+         * @return the function output
+         */
+        @Override
+        public String apply(String input, Map<String,Object> options) {
+            if (Objects.isNull(input))
+                return "";
+            if (options.containsKey("length")) {
+                String optionString = options.get("length").toString();
+                int length = 0;
+                try {
+                    length = Integer.valueOf(optionString).intValue();
+                }
+                catch (NumberFormatException nfe) {
+                    logger.error(TEMPLATE, "The '{}' option value '{}' for the '{}' function is not a valid integer", "length", optionString, NAME);
+                    return input;
+                }
+                if (length < 0) {
+                    logger.error(TEMPLATE, "The '{}' option value '{}' for the '{}' function cannot be negative", "length", optionString, NAME);
+                    return input;
+                }
+                return input.length() > length ? input.substring(input.length()-length, input.length()) : input;
+            }
+            else return input;
+        }
+    }
+
+    /**
+     * This function returns the first {@code N} characters of the input string, where {@code N} is the {@code length} option.
+     * If the input string is shorter or the same length than the parameter, the whole input string is returned unchanged.
+     * If the input is {@code null} an empty string is returned.
+     */
+    public static class CutRight extends AbstractParametricFunction {
+        /**
+         * The name to use for this function.
+         */
+        public static final String NAME = "cutRight";
+
+        /**
+         * This method returns the first {@code N} characters of the input string, where {@code N} is the {@code length} option.
+         * If the input string is shorter or the same length than the parameter, the whole input string is returned unchanged.
+         * If the input is {@code null} an empty string is returned.
+         * 
+         * @param input the function input
+         * @param options the function options, where the option named {@code length} determines the maximum length of the
+         * returned string. If there is no such option the same string given as input is returned
+         * 
+         * @return the function output
+         */
+        @Override
+        public String apply(String input, Map<String,Object> options) {
+            if (Objects.isNull(input))
+                return "";
+            if (options.containsKey("length")) {
+                String optionString = options.get("length").toString();
+                int length = 0;
+                try {
+                    length = Integer.valueOf(optionString).intValue();
+                }
+                catch (NumberFormatException nfe) {
+                    logger.error(TEMPLATE, "The '{}' option value '{}' for the '{}' function is not a valid integer", "length", optionString, NAME);
+                    return input;
+                }
+                if (length < 0) {
+                    logger.error(TEMPLATE, "The '{}' option value '{}' for the '{}' function cannot be negative", "length", optionString, NAME);
+                    return input;
+                }
+                return input.length() > length ? input.substring(0, length) : input;
+            }
+            else return input;
+        }
+    }
+
+    /**
+     * This function replaces the given character sequences from the input string with the given replacement.
+     */
+    public static class Replace extends AbstractParametricFunction {
+        /**
+         * The name to use for this function.
+         */
+        public static final String NAME = "replace";
+
+        /**
+         * This method replaces the given character sequences from the input string with the given replacement.
+         * 
+         * @param input the function input
+         * @param options the function options, where the option named {@code from} determines the character sequence to be replaced,
+         * {@code to} determines the character sequence to be used for replacement
+         * 
+         * @return the function output
+         */
+        @Override
+        public String apply(String input, Map<String,Object> options) {
+            if (Objects.isNull(input))
+                return "";
+            if (options.containsKey("from")) {
+                String fromString = options.get("from").toString();
+                String toString = "";
+                if (options.containsKey("to")) {
+                    toString = options.get("to").toString();
+                }
+                return input.replace(fromString, toString);
+            }
+            else return input;
+        }
+    }
+
+    /**
+     * This function returns a time value (expressed in milliseconds) and is also able to format it according to an optional
+     * format string.
+     */
+    public static class TimeFormat extends AbstractParametricFunction {
+        /**
+         * The name to use for this function.
+         */
+        public static final String NAME = "timeFormat";
+
+        /**
+         * This method returns returns a time value (expressed in milliseconds) and is also able to format it according to an optional
+         * format string.
+         * 
+         * @param input the function input. This must be a long value representing a timestamp in milliseconds
+         * since January 1, 1970, 00:00:00 GMT. If this value is not provided, the system current timestamp is used.
+         * @param options the function options, where the option named {@code format} determines the string representation of the timestamp
+         * 
+         * @return the function output
+         */
+        @Override
+        public String apply(String input, Map<String,Object> options) {
+            long currentTime = System.currentTimeMillis();
+            if ((!Objects.isNull(input)) && !input.isBlank()) {
+                try {
+                    currentTime = Long.valueOf(input).longValue();
+                }
+                catch (NumberFormatException nfe) {
+                    logger.error(TEMPLATE, "The value '{}' for the '{}' function is not a valid long", input, NAME);
+                    return "";
+                }
+                if (currentTime < 0) {
+                    logger.error(TEMPLATE, "The value '{}' for the '{}' function cannot be negative", input, NAME);
+                    return "";
+                }
+            }
+
+            if (options.containsKey("format")) {
+                String formatString = options.get("format").toString();
+                try {
+                    Date date = new Date(Long.valueOf(currentTime).longValue());
+                    SimpleDateFormat dateFormat = new SimpleDateFormat(formatString);
+                    dateFormat.setTimeZone(TimeZone.getTimeZone(ZoneOffset.UTC));
+                    return dateFormat.format(date);
+                }
+                catch (IllegalArgumentException iae) {
+                    logger.error(TEMPLATE, "String '{}' does not represent a valid date format string", formatString);
+                    return "";
+                }
+            }
+            else return Long.valueOf(currentTime).toString();
         }
     }
 }